pyproject.toml
[tool.poetry]
name = "mba-tcc"
version = "0.1.0"
description = ""
authors = ["Paulo Haddad <paulochf@gmail.com>"]
readme = "README.md"
packages = [{include = "mba_tcc"}]

[tool.poetry.dependencies]
python = "^3.10"
jupyter = "*"
jupyterlab = "^3.4.7"
numpy = "*"
pandas = "*"
prefect = "*"
prefect-dask = "*"
pyarrow = "*"
python-dotenv = {extras = ["cli"], version = "^0.21.0"}
#scipy = "*"
#seaborn = "*"
typer = "*"


[tool.poetry.group.dev.dependencies]
autoflake = "*"
black = "*"
devtools = "*"
ipdb = "*"
ipython = "*"
isort = "*"
pylint = "*"
pytest = "*"
pytest-mock = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


# Python-style entrypoints and scripts are easily expressed
[tool.poetry.scripts]
main = "mba_tcc.run_pipe:main"
test = "pytest"
